fix warning

+++ maps/pizza_ya_san/weapon_as_shotgun.cpp
@@ -22,7 +22,7 @@ ItemInfo g_shotgun_info = {
 	2,								// iSlot
 	-1,								// iPosition (-1 = automatic)
 	"buckshot",						// pszAmmo1
-	gSkillData.sk_ammo_max_buckshot,// iMaxAmmo1
+	(int)gSkillData.sk_ammo_max_buckshot,// iMaxAmmo1
 	NULL,							// pszAmmo2
 	-1,								// iMaxAmmo2
 	"pizza_ya_san/weapon_as_shotgun",// pszName (path to HUD config)
